upgrade to dotnet core 1.0 rc2
