+++ src/Aqua/TypeSystem/Emit/TypeEmitter.cs
@@ -1,6 +1,6 @@
 ﻿// Copyright (c) Christof Senn. All rights reserved. See license.txt in the project root for license information.
 
-#if NET || NET35 || CORECLR
+#if NET || NET35 || NETSTANDARD || CORECLR
 
 namespace Aqua.TypeSystem.Emit
 {
